refactor: type and function to under src
