+++ src/types/property.ts
